requirements.txt
# Requirements for boardcast_device.py
paho-mqtt==1.6.1             # MQTT client library
python-dotenv==1.0.0          # .env file loader
pynmea2==1.18.0               # NMEA GPS parser
pyserial==3.5                 # Serial communication
smbus2==0.4.2                 # I2C communication
RPi.bme280==0.2.3             # BME280 sensor driver